Move NLopt test to separate
