+++ test/Project.toml
@@ -2,12 +2,9 @@
 ArrayDiff = "c45fa1ca-6901-44ac-ae5b-5513a4852d50"
 Calculus = "49dc2e85-a5d0-5ad3-a950-438e2897f1b9"
 GenOpt = "f2c049d8-7489-4223-990c-4f1c121a4cde"
-Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-NLPModelsIpopt = "f4238b75-b362-5c4c-b852-0801c9a21d71"
-NLPModelsJuMP = "792afdf1-32c1-5681-94e0-d7bf7a5df49e"
 NLopt = "76087f3c-5699-56af-9a33-bf431cd00edd"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
