+++ test/runtests.jl
@@ -1,3 +1,5 @@
 include("ReverseAD.jl")
 include("ArrayDiff.jl")
 include("JuMP.jl")
+# Currently needs https://github.com/jump-dev/NLopt.jl/pull/273
+#include("NLopt.jl")
